+++ library/src/schemas/optional/optional.test-d.ts
@@ -9,16 +9,10 @@ import { optional, type OptionalSchema } from './optional.ts';
 
 describe('optional', () => {
   describe('should return schema object', () => {
-    test('with never default', () => {
-      expectTypeOf(optional(string())).toEqualTypeOf<
-        OptionalSchema<StringSchema<undefined>, never>
-      >();
-    });
-
     test('with undefined default', () => {
-      expectTypeOf(optional(string(), undefined)).toEqualTypeOf<
-        OptionalSchema<StringSchema<undefined>, undefined>
-      >();
+      type Schema = OptionalSchema<StringSchema<undefined>, undefined>;
+      expectTypeOf(optional(string())).toEqualTypeOf<Schema>();
+      expectTypeOf(optional(string(), undefined)).toEqualTypeOf<Schema>();
     });
 
     test('with undefined getter default', () => {
@@ -41,11 +35,10 @@ describe('optional', () => {
   });
 
   describe('should infer correct types', () => {
-    type Schema1 = OptionalSchema<StringSchema<undefined>, never>;
-    type Schema2 = OptionalSchema<StringSchema<undefined>, undefined>;
-    type Schema3 = OptionalSchema<StringSchema<undefined>, 'foo'>;
-    type Schema4 = OptionalSchema<StringSchema<undefined>, () => undefined>;
-    type Schema5 = OptionalSchema<StringSchema<undefined>, () => 'foo'>;
+    type Schema1 = OptionalSchema<StringSchema<undefined>, undefined>;
+    type Schema2 = OptionalSchema<StringSchema<undefined>, 'foo'>;
+    type Schema3 = OptionalSchema<StringSchema<undefined>, () => undefined>;
+    type Schema4 = OptionalSchema<StringSchema<undefined>, () => 'foo'>;
 
     test('of input', () => {
       type Input = string | undefined;
@@ -53,15 +46,13 @@ describe('optional', () => {
       expectTypeOf<InferInput<Schema2>>().toEqualTypeOf<Input>();
       expectTypeOf<InferInput<Schema3>>().toEqualTypeOf<Input>();
       expectTypeOf<InferInput<Schema4>>().toEqualTypeOf<Input>();
-      expectTypeOf<InferInput<Schema5>>().toEqualTypeOf<Input>();
     });
 
     test('of output', () => {
       expectTypeOf<InferOutput<Schema1>>().toEqualTypeOf<string | undefined>();
-      expectTypeOf<InferOutput<Schema2>>().toEqualTypeOf<string | undefined>();
-      expectTypeOf<InferOutput<Schema3>>().toEqualTypeOf<string>();
-      expectTypeOf<InferOutput<Schema4>>().toEqualTypeOf<string | undefined>();
-      expectTypeOf<InferOutput<Schema5>>().toEqualTypeOf<string>();
+      expectTypeOf<InferOutput<Schema2>>().toEqualTypeOf<string>();
+      expectTypeOf<InferOutput<Schema3>>().toEqualTypeOf<string | undefined>();
+      expectTypeOf<InferOutput<Schema4>>().toEqualTypeOf<string>();
     });
 
     test('of issue', () => {
@@ -69,7 +60,6 @@ describe('optional', () => {
       expectTypeOf<InferIssue<Schema2>>().toEqualTypeOf<StringIssue>();
       expectTypeOf<InferIssue<Schema3>>().toEqualTypeOf<StringIssue>();
       expectTypeOf<InferIssue<Schema4>>().toEqualTypeOf<StringIssue>();
-      expectTypeOf<InferIssue<Schema5>>().toEqualTypeOf<StringIssue>();
     });
   });
 });