+++ library/src/types/object.ts
@@ -103,11 +103,8 @@ type QuestionMarkSchema =
 /**
  * Has default type.
  */
-type HasDefault<TSchema extends QuestionMarkSchema> = [
-  TSchema['default'],
-] extends [never]
-  ? false
-  : true;
+type HasDefault<TSchema extends QuestionMarkSchema> =
+  undefined extends TSchema['default'] ? false : true;
 
 /**
  * Exact optional input type.
@@ -117,8 +114,8 @@ type ExactOptionalInput<
     | BaseSchema<unknown, unknown, BaseIssue<unknown>>
     | BaseSchemaAsync<unknown, unknown, BaseIssue<unknown>>,
 > = TSchema extends
-  | OptionalSchema<infer TWrapped, never>
-  | OptionalSchemaAsync<infer TWrapped, never>
+  | OptionalSchema<infer TWrapped, unknown>
+  | OptionalSchemaAsync<infer TWrapped, unknown>
   ? ExactOptionalInput<TWrapped>
   : InferInput<TSchema>;
 
@@ -130,8 +127,8 @@ type ExactOptionalOutput<
     | BaseSchema<unknown, unknown, BaseIssue<unknown>>
     | BaseSchemaAsync<unknown, unknown, BaseIssue<unknown>>,
 > = TSchema extends
-  | OptionalSchema<infer TWrapped, never>
-  | OptionalSchemaAsync<infer TWrapped, never>
+  | OptionalSchema<infer TWrapped, unknown>
+  | OptionalSchemaAsync<infer TWrapped, unknown>
   ? HasDefault<TSchema> extends true
     ? InferOutput<TSchema>
     : ExactOptionalOutput<TWrapped>
@@ -178,7 +175,7 @@ type OptionalOutputKeys<TEntries extends ObjectEntries | ObjectEntriesAsync> = {
  */
 type InputWithQuestionMarks<
   TEntries extends ObjectEntries | ObjectEntriesAsync,
-  TObject extends InferEntriesInput<TEntries> | InferEntriesOutput<TEntries>,
+  TObject extends InferEntriesInput<TEntries>,
 > = MarkOptional<TObject, OptionalInputKeys<TEntries>>;
 
 /**
@@ -186,7 +183,7 @@ type InputWithQuestionMarks<
  */
 type OutputWithQuestionMarks<
   TEntries extends ObjectEntries | ObjectEntriesAsync,
-  TObject extends InferEntriesInput<TEntries> | InferEntriesOutput<TEntries>,
+  TObject extends InferEntriesOutput<TEntries>,
 > = MarkOptional<TObject, OptionalOutputKeys<TEntries>>;
 
 /**