+++ library/CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to the library will be documented in this file.
 - Add `args` and `returns` action to transform functions (issue #243)
 - Change types and implementation to support Standard Schema
 - Change behaviour of `minValue` and `maxValue` for `NaN` (pull request #843)
+- Change type and behaviour of `nullable`, `nullableAsync`, `nullish`, `nullishAsync`, `optional`, `optionalAsync`, `undefinedable` and `undefinedableAsync` for undefined default value (issue #878)
 - Refactor `bytes`, `maxBytes`, `minBytes` and `notBytes` action
 
 ## v0.42.1 (September 20, 2024)